Update README.md
Choose a model section in prediction has inaccurate information about the resolution for Clathrin-Coated Pits and Nuclear Pores. Originally 5x5x5, actually 4x4x4 as seen in the database for those experiments with the IDs 1986 and 2000. Therefore, we should combine them with the previous bullet point (high pressure freezing and 4x4x4).